Update and organize files
Changes made:
 - Move components for home and navigation to separate folders
 - Updated home screen
