+++ components/home/DayCarousel.tsx
